Refactor: 프로필페이지와 마이페이지에서 공통으로 쓰이는 사용자 정보 보여주는 section 컴포넌트화 작업 및 그에 따른 마이페이지 수정
